+++ CHANGELOG.md
@@ -1,8 +1,14 @@
-## 1.5.0
+## 1.5.x
 
 Features:
 
  - "Extension updated" links to the changelog
- - Download current url when no link or text selected
+ - Download current page's url when no link or text selected
  - Notification when reached 75% of quota for limited hosters
- - Notification when less than 7 days left of premium
+ - Notification when only 7 days left of premium
+
+Bugfixes:
+
+ - Persistent storage of warnings
+ - Handle case where account information is requested when not logged in
+ - No "extension updated" notification in case of bugfix